requirements.txt
Django==2.2.25
django-cors-headers==3.10.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
pkg_resources==0.0.0
postgres==4.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
psycopg2-pool==1.1
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
uuid==1.30
